azure-ai/computer-vision/requirements.txt
python-dotenv
azure-cognitiveservices-vision-computervision 
opencv-python 
pillow 
numpy 
msrest
gradio